feat(files): make imgproxy the default image resizer, remove the ff
Drop the imgproxy_image_resize flag and the ConvertAPI resize path now
that imgproxy handles all raster image resizing. The two-provider
ImageConverter abstraction collapses to a single concrete class, and the
flag branch in images.ts goes away.

ConvertAPI is still used for file-format conversion in file_generation,
so its key stays — routed through config.getConvertAPIKey() instead of
process.env.
