+++ front/types/shared/feature_flags.ts
@@ -345,11 +345,6 @@ export const WHITELISTABLE_FEATURES_CONFIG = {
       "Per-pod default agent: pre-select an agent for new conversations started in a project (pod).",
     stage: "dust_only",
   },
-  imgproxy_image_resize: {
-    description:
-      "Resize uploaded raster images via imgproxy instead of ConvertAPI",
-    stage: "dust_only",
-  },
   workspace_default_agent: {
     description:
       "Workspace default agent: admins can pre-select a workspace-wide default agent for new conversations.",
